pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bobr"
version = "0.1.0"
description = "Git-native backlog management for AI-assisted teams"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "python-frontmatter>=1.1",
    "rich>=13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.4",
]

[project.scripts]
bobr = "bobr.cli.main:entry"

[tool.hatch.build.targets.wheel]
packages = ["src/bobr"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
